drop swift-package-factory dependency
