+++ Sources/BSONDecoding/Decoder/BSON.KeyedDecoder.Super.swift
@@ -1,0 +1,7 @@
+extension BSON.KeyedDecoder
+{
+    enum Super:String, CodingKey 
+    {
+        case `super`
+    }
+}
