--- Sources/BSONDecoding/Decoder/Decoder.spf
@@ -1,172 +1,0 @@
-import BSONUnions
-@basis 
-let nonthrowing:[Never] = 
-[
-    Bool,
-    Float,
-    Double,
-    String
-]
-@basis 
-let throwing:[Never] = 
-[
-    Int,
-    Int64,
-    Int32,
-    Int16,
-    Int8,
-    UInt,
-    UInt64,
-    UInt32,
-    UInt16,
-    UInt8,
-]
-
-extension BSON.Decoder:SingleValueDecodingContainer
-{
-    public 
-    func decode<T>(_:T.Type) throws -> T where T:Decodable
-    {
-        try .init(from: self)
-    }
-    public
-    func decodeNil() -> Bool
-    {
-        self.value.null != nil
-    }
-    @matrix(__Primitive__: nonthrowing)
-    public
-    func decode(_:__Primitive__.Type) throws -> __Primitive__
-    {
-        try self.diagnose { $0.as(__Primitive__.self) }
-    }
-    @matrix(__Primitive__: throwing)
-    public
-    func decode(_:__Primitive__.Type) throws -> __Primitive__
-    {
-        try self.diagnose { try $0.as(__Primitive__.self) }
-    }
-}
-
-extension BSON.KeyedDecoder:KeyedDecodingContainerProtocol 
-{
-    public
-    func decode<T>(_:T.Type, forKey key:Key) throws -> T where T:Decodable
-    {
-        return try .init(from: try self.singleValueContainer(forKey: key))
-    }
-    func decodeNil(forKey key:Key) throws -> Bool
-    {
-        try self.diagnose(key) { $0.null != nil }
-    }
-    @matrix(__Primitive__: nonthrowing)
-    public
-    func decode(_:__Primitive__.Type, forKey key:Key) throws -> __Primitive__
-    {
-        try self.diagnose(key) { $0.as(__Primitive__.self) }
-    }
-    @matrix(__Primitive__: throwing)
-    public
-    func decode(_:__Primitive__.Type, forKey key:Key) throws -> __Primitive__
-    {
-        try self.diagnose(key) { try $0.as(__Primitive__.self) }
-    }
-    
-    func superDecoder() throws -> any Decoder
-    {
-        try self.singleValueContainer(forKey: BSON.ObjectKey.super, typed: BSON.ObjectKey.self)
-    }
-    public 
-    func superDecoder(forKey key:Key) throws -> any Decoder
-    {
-        try self.singleValueContainer(forKey: key) as any Decoder
-    }
-    
-    public 
-    func singleValueContainer<Key>(forKey key:Key,
-        typed _:Key.Type = Key.self) throws -> BSON.Decoder<Bytes>
-        where Key:CodingKey
-    {
-        let value:AnyBSON<Bytes> = try self.diagnose(key){ $0 }
-        let decoder:BSON.Decoder<Bytes> = .init(value, 
-            path: self.codingPath + CollectionOfOne<any CodingKey>.init(key))
-        return decoder
-    }
-    public 
-    func nestedUnkeyedContainer(forKey key:Key) throws -> UnkeyedDecodingContainer
-    {
-        let path:[any CodingKey] = self.codingPath + CollectionOfOne<any CodingKey>.init(key)
-        let container:BSON.UnkeyedDecoder<Bytes.SubSequence> =
-            .init(try self.diagnose(key) { try $0.array() }, path: path)
-        return container as UnkeyedDecodingContainer
-    }
-    public 
-    func nestedContainer<NestedKey>(keyedBy _:NestedKey.Type,
-        forKey key:Key) throws -> KeyedDecodingContainer<NestedKey>
-    {
-        let path:[any CodingKey] = self.codingPath + CollectionOfOne<any CodingKey>.init(key)
-        let container:BSON.KeyedDecoder<Bytes.SubSequence, NestedKey> =
-            .init(try self.diagnose(key) { try $0.dictionary() }, path: path)
-        return .init(container)
-    }
-}
-
-extension BSON.UnkeyedDecoder:UnkeyedDecodingContainer
-{
-    public mutating 
-    func decode<T>(_:T.Type) throws -> T where T:Decodable
-    {
-        try .init(from: try self.singleValueContainer())
-    }
-    public mutating 
-    func decodeNil() throws -> Bool
-    {
-        try self.diagnose { $0.null != nil }
-    }
-    @matrix(__Primitive__: nonthrowing)
-    public mutating 
-    func decode(_:__Primitive__.Type) throws -> __Primitive__
-    {
-        try self.diagnose { $0.as(__Primitive__.self) }
-    }
-    @matrix(__Primitive__: throwing)
-    public mutating 
-    func decode(_:__Primitive__.Type) throws -> __Primitive__
-    {
-        try self.diagnose { try $0.as(__Primitive__.self) }
-    }
-    
-    public mutating  
-    func superDecoder() throws -> any Decoder
-    {
-        try self.singleValueContainer() as any Decoder
-    }
-    public mutating 
-    func singleValueContainer() throws -> BSON.Decoder<Bytes>
-    {
-        let key:BSON.ListKey = .init(intValue: self.currentIndex) 
-        let value:AnyBSON<Bytes> = try self.diagnose { $0 }
-        let decoder:BSON.Decoder<Bytes> = .init(value, 
-            path: self.codingPath + CollectionOfOne<any CodingKey>.init(key))
-        return decoder
-    }
-    public mutating 
-    func nestedUnkeyedContainer() throws -> any UnkeyedDecodingContainer
-    {
-        let path:[any CodingKey] = self.codingPath +
-            CollectionOfOne<any CodingKey>.init(BSON.ListKey.init(intValue: self.currentIndex))
-        let container:BSON.UnkeyedDecoder<Bytes.SubSequence> =
-            .init(try self.diagnose { try $0.array() }, path: path)
-        return container as any UnkeyedDecodingContainer
-    }
-    public mutating 
-    func nestedContainer<NestedKey>(keyedBy _:NestedKey.Type) 
-        throws -> KeyedDecodingContainer<NestedKey>
-    {
-        let path:[any CodingKey] = self.codingPath + 
-            CollectionOfOne<any CodingKey>.init(BSON.ListKey.init(intValue: self.currentIndex))
-        let container:BSON.KeyedDecoder<Bytes.SubSequence, NestedKey> = 
-            .init(try self.diagnose { try $0.dictionary() }, path: path)
-        return .init(container)
-    }
-}
